add 3 controller StaticPagesController, UsersControlller, SessionsController, add 3 routes ‘/’, ‘signup’, ‘login’, add Models folders and move the User.php on the Models
