+++ app/Http/Controllers/UsersController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class UsersController extends Controller
+{
+    // 返回用户注册表单页面
+    public function create()
+    {
+        return '注册页面';
+    }
+}
